% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{add}
\alias{add}
\title{Title a function to paste string list like paste0}
\usage{
add(list)
}
\arguments{
\item{list}{the  list of string that you want to paste}
}
\value{
a pasted string
}
\description{
Title a function to paste string list like paste0
}
\examples{
library(goodpkg)
print(add(c("this ", "function ", "is ok")))
}
